Rename from "noflo-gegl" to "imgflo"
The project and code is independent from NoFlo, and for those who do not know GEGL
the relationship with image and graphics processing was not clear.
Hopefully new name makes these things clear.
imgflo is similar to NoFlo and MicroFlo as it is another runtime that
one can target using the Flowhub.io visual-programming IDE.