server: Allow to parametrize processing requests on HTTP API
The graphs currently have to reside in the "examples" directory,
and should export inports/outports for all the properties that can be configured.
In the HTTP API, one refers to these properties using the query string.
The special ports named "input" and "output" are mandatory, and is
where the image data will flow through the subgraph. The server
uses these to attach file loading and saving operations.